pyproject.toml
[project]
name = "astro-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.6"
dependencies = [
    "gradio>=6.2.0",
    "matplotlib>=3.10.8",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.20",
    "pillow>=11.0.0",
]
